Move File-based APIs to extensions
